Fixed more lint errors
