+++ stylecop.ruleset
@@ -6,10 +6,8 @@
   <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
     <Rule Id="SA1101" Action="None" /> <!-- Prefix local calls with this -->
     <Rule Id="SA1200" Action="None" /> <!-- Using clauses inside namespaces -->
+    <Rule Id="SA1502" Action="None" /> <!-- Single-lined elements -->
 
-    <Rule Id="SA1402" Action="None" />
-    <Rule Id="SA1502" Action="None" />
-    <Rule Id="SA1513" Action="None" />
     <Rule Id="SA1515" Action="None" />
     <Rule Id="SA1600" Action="None" />
     <Rule Id="SA1611" Action="None" />
